src/music-bot/requirements.txt
discord.py
mpv
